+++ code/modules/robotics/bot/buttbot.dm
@@ -3,7 +3,7 @@
 /obj/machinery/bot/buttbot
 	name = "buttbot"
 	desc = "Well I... uh... huh."
-	icon = 'icons/obj/aibots.dmi'
+	icon = 'icons/obj/bots/aibots.dmi'
 	icon_state = "buttbot"
 	layer = 5.0 // Todo layer
 	density = 0
@@ -51,7 +51,7 @@
 		if (user)
 			user.show_text("You short out the vocal emitter on [src].", "red")
 		SPAWN_DBG(0)
-			src.visible_message("<span style=\"color:red\"><B>[src] buzzes oddly!</B></span>")
+			src.visible_message("<span class='alert'><B>[src] buzzes oddly!</B></span>")
 			playsound(src.loc, "sound/misc/extreme_ass.ogg", 50, 1)
 		src.emagged = 1
 		return 1
@@ -69,13 +69,8 @@
 	if (istype(W, /obj/item/card/emag))
 		//Do not hit the buttbot with the emag tia
 	else
-		src.visible_message("<span style=\"color:red\">[user] hits [src] with [W]!</span>")
-		switch(W.damtype)
-			if("fire")
-				src.health -= W.force * 0.5
-			if("brute")
-				src.health -= W.force * 0.5
-			else
+		src.visible_message("<span class='alert'>[user] hits [src] with [W]!</span>")
+		src.health -= W.force * 0.5
 		if (src.health <= 0)
 			src.explode()
 
@@ -104,10 +99,10 @@
 	return src.explode()
 
 /obj/machinery/bot/buttbot/explode()
+	if(src.exploding) return
+	src.exploding = 1
 	src.on = 0
-	src.visible_message("<span style=\"color:red\"><B>[src] blows apart!</B></span>")
-	var/datum/effects/system/spark_spread/s = unpool(/datum/effects/system/spark_spread)
-	s.set_up(3, 1, src)
-	s.start()
+	src.visible_message("<span class='alert'><B>[src] blows apart!</B></span>")
+	elecflash(src, radius=1, power=3, exclude_center = 0)
 	qdel(src)
 	return
